Working easyBot with extra stuff
Working bot that guesses the same number only once.
Made it so the player and the bot has each their turn.
Bot's turn automatically ends after 2 seconds.
When the bot turn's start the input field is disabled and a message display showing it's the bot's turn.
When it is the player's time to play a message will display notifying that it's their turn.