Fixed a few bugs
Made it so the timer will stop as soon as you leave the game page.
The timer will also stop as soon as you enter the Winner page.
Now the game won show the wrong messages at the wrong times or suddenly take you to the Winner page while navigating other pages.