Small details
Cleaned the code a bit.
Updated some of the CSS to make the page cleaner.
Made it so the input can have MAX 4 digits.
Made it the input has a placeholder.
Made it so if the Game page is refreshed it will redirect to the home menu.
Updated the way the "Guess Higher" and "Guess Low" alert shows.